[#112] refactor : paperTheme Constants로 분리 - 메인 페이지에 재사용 예정
